Finish GMMHMM tests
