+++ lib/sequentia/classifiers/__init__.py
@@ -1,5 +1,5 @@
 from .hmm import (
-    HMM, GMMHMM, HMMClassifier,
+    GMMHMM, HMMClassifier,
     _Topology, _LeftRightTopology, _ErgodicTopology, _StrictLeftRightTopology
 )
 from .knn import KNNClassifier
