+++ lib/sequentia/classifiers/hmm/__init__.py
@@ -1,4 +1,3 @@
-from .hmm import HMM
 from .gmmhmm import GMMHMM
 from .hmm_classifier import HMMClassifier
 from .topologies import _Topology, _LeftRightTopology, _ErgodicTopology, _StrictLeftRightTopology
