+++ lib/test/lib/classifiers/hmm/test_gmmhmm.py
@@ -1,9 +1,6 @@
-import pytest, warnings, os, json, numpy as np
+import pytest, warnings, os, numpy as np, hmmlearn.hmm
 from copy import deepcopy
-with warnings.catch_warnings():
-    warnings.filterwarnings('ignore', category=DeprecationWarning)
-    import pomegranate as pg
-from sequentia.classifiers import HMM, GMMHMM, _LeftRightTopology, _ErgodicTopology, _StrictLeftRightTopology
+from sequentia.classifiers import GMMHMM, _LeftRightTopology, _ErgodicTopology, _StrictLeftRightTopology
 from ....support import assert_equal, assert_not_equal
 
 # Set seed for reproducible randomness
@@ -16,201 +13,618 @@ X = [rng.random((10 * i, 3)) for i in range(1, 4)]
 x = rng.random((15, 3))
 
 # Unparameterized HMMs
-hmm_diag = GMMHMM(label='c1', n_states=5, n_components=5, covariance='diagonal', random_state=rng)
-hmm_full = GMMHMM(label='c1', n_states=5, n_components=5, covariance='full', random_state=rng)
+hmm_lr = GMMHMM(label='c1', n_states=5, topology='left-right', random_state=rng)
+hmm_e = GMMHMM(label='c1', n_states=5, topology='ergodic', random_state=rng)
+hmm_slr = GMMHMM(label='c1', n_states=5, topology='strict-left-right', random_state=rng)
+
+# ======================================================== #
+# GMMHMM.set_uniform_initial() + GMMHMM.initial (property) #
+# ======================================================== #
+
+def test_left_right_uniform_initial():
+    """Uniform initial state distribution for a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_uniform_initial()
+    assert_equal(hmm.initial, np.array([
+        0.2, 0.2, 0.2, 0.2, 0.2
+    ]))
+
+def test_ergodic_uniform_initial():
+    """Uniform initial state distribution for an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    hmm.set_uniform_initial()
+    assert_equal(hmm.initial, np.array([
+        0.2, 0.2, 0.2, 0.2, 0.2
+    ]))
+
+def test_strict_left_right_uniform_initial():
+    """Uniform initial state distribution for a strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    hmm.set_uniform_initial()
+    assert_equal(hmm.initial, np.array([
+        0.2, 0.2, 0.2, 0.2, 0.2
+    ]))
+
+# ======================================================= #
+# GMMHMM.set_random_initial() + GMMHMM.initial (property) #
+# ======================================================= #
+
+def test_left_right_random_initial():
+    """Random initial state distribution for a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_random_initial()
+    assert_equal(hmm.initial, np.array([
+        0.35029635, 0.13344569, 0.02784745, 0.33782453, 0.15058597
+    ]))
+
+def test_ergodic_random_initial():
+    """Random initial state distribution for an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    hmm.set_random_initial()
+    assert_equal(hmm.initial, np.array([
+        0.35029635, 0.13344569, 0.02784745, 0.33782453, 0.15058597
+    ]))
+
+def test_strict_left_right_random_initial():
+    """Random initial state distribution for a strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    hmm.set_random_initial()
+    assert_equal(hmm.initial, np.array([
+        0.35029635, 0.13344569, 0.02784745, 0.33782453, 0.15058597
+    ]))
+
+# ================================================================ #
+# GMMHMM.set_uniform_transitions() + GMMHMM.transitions (property) #
+# ================================================================ #
+
+def test_left_right_uniform_transitions():
+    """Uniform transition matrix for a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_uniform_transitions()
+    assert_equal(hmm.transitions, np.array([
+        [0.2, 0.2 , 0.2       , 0.2       , 0.2       ],
+        [0. , 0.25, 0.25      , 0.25      , 0.25      ],
+        [0. , 0.  , 0.33333333, 0.33333333, 0.33333333],
+        [0. , 0.  , 0.        , 0.5       , 0.5       ],
+        [0. , 0.  , 0.        , 0.        , 1.        ]
+    ]))
+
+def test_ergodic_uniform_transitions():
+    """Uniform transition matrix for an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    hmm.set_uniform_transitions()
+    assert_equal(hmm.transitions, np.array([
+        [0.2, 0.2, 0.2, 0.2, 0.2],
+        [0.2, 0.2, 0.2, 0.2, 0.2],
+        [0.2, 0.2, 0.2, 0.2, 0.2],
+        [0.2, 0.2, 0.2, 0.2, 0.2],
+        [0.2, 0.2, 0.2, 0.2, 0.2]
+    ]))
+
+def test_strict_left_right_uniform_transitions():
+    """Uniform transition matrix for a strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    hmm.set_uniform_transitions()
+    assert_equal(hmm.transitions, np.array([
+        [0.5, 0.5, 0. , 0. , 0. ],
+        [0. , 0.5, 0.5, 0. , 0. ],
+        [0. , 0. , 0.5, 0.5, 0. ],
+        [0. , 0. , 0. , 0.5, 0.5],
+        [0. , 0. , 0. , 0. , 1. ]
+    ]))
+
+# =============================================================== #
+# GMMHMM.set_random_transitions() + GMMHMM.transitions (property) #
+# =============================================================== #
+
+def test_left_right_random_transitions():
+    """Random transition matrix for a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_random_transitions()
+    assert_equal(hmm.transitions, np.array([
+        [0.35029635, 0.13344569, 0.02784745, 0.33782453, 0.15058597],
+        [0.        , 0.20580715, 0.5280311 , 0.06521685, 0.20094491],
+        [0.        , 0.        , 0.33761567, 0.2333124 , 0.42907193],
+        [0.        , 0.        , 0.        , 0.36824778, 0.63175222],
+        [0.        , 0.        , 0.        , 0.        , 1.        ]
+    ]))
+
+def test_ergodic_random_transitions():
+    """Random transition matrix for an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    hmm.set_random_transitions()
+    assert_equal(hmm.transitions, np.array([
+        [0.35029635, 0.13344569, 0.02784745, 0.33782453, 0.15058597],
+        [0.19252534, 0.15767581, 0.47989976, 0.01708551, 0.15281357],
+        [0.19375092, 0.16425506, 0.21828034, 0.11397708, 0.30973661],
+        [0.46906977, 0.02941216, 0.17137502, 0.0333193 , 0.29682374],
+        [0.21312406, 0.35221103, 0.08556524, 0.06613143, 0.28296824]
+    ]))
+
+def test_strict_left_right_random_transitions():
+    """Random transition matrix for a strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    hmm.set_random_transitions()
+    assert_equal(hmm.transitions, np.array([
+        [0.72413873, 0.27586127, 0.        , 0.        , 0.        ],
+        [0.        , 0.07615418, 0.92384582, 0.        , 0.        ],
+        [0.        , 0.        , 0.81752797, 0.18247203, 0.        ],
+        [0.        , 0.        , 0.        , 0.24730529, 0.75269471],
+        [0.        , 0.        , 0.        , 0.        , 1.        ]
+    ]))
+
+# ============================ #
+# GMMHMM.fit() + GMMHMM.n_seqs #
+# ============================ #
+
+def test_fit_without_initial_and_transition():
+    """Fitting before setting the initial state distribution and transition matrix"""
+    hmm = deepcopy(hmm_lr)
+    with pytest.raises(AttributeError) as e:
+        hmm.fit(X)
+    assert str(e.value) == 'Must specify initial state distribution and transitions before the HMM can be fitted'
+
+def test_fit_without_transitions():
+    """Fitting before setting the initial transition matrix"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_uniform_initial()
+    with pytest.raises(AttributeError) as e:
+        hmm.fit(X)
+    assert str(e.value) == 'Must specify initial state distribution and transitions before the HMM can be fitted'
+
+def test_fit_without_initial():
+    """Fitting before setting the initial state distribution"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_uniform_transitions()
+    with pytest.raises(AttributeError) as e:
+        hmm.fit(X)
+    assert str(e.value) == 'Must specify initial state distribution and transitions before the HMM can be fitted'
+
+def test_fit_sets_internals():
+    """Check that fitting sets internal attributes"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_uniform_initial()
+    hmm.set_uniform_transitions()
+    hmm.fit(X)
+    assert hmm.n_seqs == 3
+    assert isinstance(hmm._model, hmmlearn.hmm.GMMHMM)
+
+def test_left_right_fit_updates_uniform_initial():
+    """Check fitting with a uniform initial state distribution of a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_uniform_initial()
+    hmm.set_uniform_transitions()
+    assert_equal(hmm.initial, np.array([
+        0.2, 0.2, 0.2, 0.2, 0.2
+    ]))
+    before = hmm.initial
+    hmm.fit(X)
+    assert_not_equal(hmm.initial, before)
+
+def test_left_right_fit_updates_random_initial():
+    """Check fitting with a random initial state distribution of a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    assert_equal(hmm.initial, np.array([
+        0.35029635, 0.13344569, 0.02784745, 0.33782453, 0.15058597
+    ]))
+    before = hmm.initial
+    hmm.fit(X)
+    assert_not_equal(hmm.initial, before)
+
+def test_left_right_fit_updates_uniform_transitions():
+    """Check fitting with a uniform transition matrix of a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_uniform_initial()
+    hmm.set_uniform_transitions()
+    before = hmm.transitions
+    assert_equal(hmm.transitions, np.array([
+        [0.2       , 0.2       , 0.2       , 0.2       , 0.2       ],
+        [0.        , 0.25      , 0.25      , 0.25      , 0.25      ],
+        [0.        , 0.        , 0.33333333, 0.33333333, 0.33333333],
+        [0.        , 0.        , 0.        , 0.5       , 0.5       ],
+        [0.        , 0.        , 0.        , 0.        , 1.        ]
+    ]))
+    hmm.fit(X)
+    assert_not_equal(hmm.transitions, before)
+
+def test_left_right_fit_updates_random_transitions():
+    """Check fitting with a random transition matrix of a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    before = hmm.transitions
+    assert_equal(hmm.transitions, np.array([
+        [0.19252534, 0.15767581, 0.47989976, 0.01708551, 0.15281357],
+        [0.        , 0.21269278, 0.26671807, 0.16241481, 0.35817434],
+        [0.        , 0.        , 0.33753566, 0.19947995, 0.46298439],
+        [0.        , 0.        , 0.        , 0.39158159, 0.60841841],
+        [0.        , 0.        , 0.        , 0.        , 1.        ]
+    ]))
+    hmm.fit(X)
+    assert_not_equal(hmm.transitions, before)
+
+def test_ergodic_fit_updates_uniform_initial():
+    """Check fitting with a uniform initial state distribution of an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    hmm.set_uniform_initial()
+    hmm.set_uniform_transitions()
+    assert_equal(hmm.initial, np.array([
+        0.2, 0.2, 0.2, 0.2, 0.2
+    ]))
+    before = hmm.initial
+    hmm.fit(X)
+    assert_not_equal(hmm.initial, before)
+
+def test_ergodic_fit_updates_random_initial():
+    """Check fitting with a random initial state distribution of an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    assert_equal(hmm.initial, np.array([
+        0.35029635, 0.13344569, 0.02784745, 0.33782453, 0.15058597
+    ]))
+    before = hmm.initial
+    hmm.fit(X)
+    assert_not_equal(hmm.initial, before)
+
+def test_ergodic_fit_updates_uniform_transitions():
+    """Check fitting with a uniform transition matrix of an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    hmm.set_uniform_initial()
+    hmm.set_uniform_transitions()
+    assert_equal(hmm.transitions, np.array([
+        [0.2, 0.2, 0.2, 0.2, 0.2],
+        [0.2, 0.2, 0.2, 0.2, 0.2],
+        [0.2, 0.2, 0.2, 0.2, 0.2],
+        [0.2, 0.2, 0.2, 0.2, 0.2],
+        [0.2, 0.2, 0.2, 0.2, 0.2]
+    ]))
+    before = hmm.transitions
+    hmm.fit(X)
+    assert_not_equal(hmm.transitions, before)
+
+def test_ergodic_fit_updates_random_transitions():
+    """Check fitting with a random transition matrix of an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    assert_equal(hmm.transitions, np.array([
+        [0.19252534, 0.15767581, 0.47989976, 0.01708551, 0.15281357],
+        [0.19375092, 0.16425506, 0.21828034, 0.11397708, 0.30973661],
+        [0.46906977, 0.02941216, 0.17137502, 0.0333193 , 0.29682374],
+        [0.21312406, 0.35221103, 0.08556524, 0.06613143, 0.28296824],
+        [0.05212313, 0.3345513 , 0.17192948, 0.16379392, 0.27760217]]))
+    before = hmm.transitions
+    hmm.fit(X)
+    assert_not_equal(hmm.transitions, before)
+
+def test_strict_left_right_fit_updates_uniform_initial():
+    """Check fitting with a uniform initial state distribution of a strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    hmm.set_uniform_initial()
+    hmm.set_uniform_transitions()
+    assert_equal(hmm.initial, np.array([
+        0.2, 0.2, 0.2, 0.2, 0.2
+    ]))
+    before = hmm.initial
+    hmm.fit(X)
+    assert_not_equal(hmm.initial, before)
+
+def test_strict_left_right_fit_updates_random_initial():
+    """Check fitting with a random initial state distribution of a strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    assert_equal(hmm.initial, np.array([
+        0.35029635, 0.13344569, 0.02784745, 0.33782453, 0.15058597
+    ]))
+    before = hmm.initial
+    hmm.fit(X)
+    assert_not_equal(hmm.initial, before)
+
+def test_strict_left_right_fit_updates_uniform_transitions():
+    """Check fitting with a uniform transition matrix of a strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    hmm.set_uniform_initial()
+    hmm.set_uniform_transitions()
+    before = hmm.transitions
+    assert_equal(hmm.transitions, np.array([
+        [0.5, 0.5, 0. , 0. , 0. ],
+        [0. , 0.5, 0.5, 0. , 0. ],
+        [0. , 0. , 0.5, 0.5, 0. ],
+        [0. , 0. , 0. , 0.5, 0.5],
+        [0. , 0. , 0. , 0. , 1. ]
+    ]))
+    hmm.fit(X)
+    assert_not_equal(hmm.transitions, before)
+
+def test_strict_left_right_fit_updates_random_transitions():
+    """Check fitting with a random transition matrix of a strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    before = hmm.transitions
+    assert_equal(hmm.transitions, np.array([
+        [0.54975645, 0.45024355, 0.        , 0.        , 0.        ],
+        [0.        , 0.96562169, 0.03437831, 0.        , 0.        ],
+        [0.        , 0.        , 0.29607315, 0.70392685, 0.        ],
+        [0.        , 0.        , 0.        , 0.42938524, 0.57061476],
+        [0.        , 0.        , 0.        , 0.        , 1.        ]
+    ]))
+    hmm.fit(X)
+    assert_not_equal(hmm.transitions, before)
+
+# ================ #
+# GMMHMM.forward() #
+# ================ #
+
+def test_forward_without_fit():
+    """Forward algorithm without fitting the HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    with pytest.raises(AttributeError) as e:
+        hmm.forward(x)
+    assert str(e.value) == 'The model must be fitted before running the forward algorithm'
+
+def test_left_right_forward():
+    """Forward algorithm on a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    hmm.fit(X)
+    assert isinstance(hmm.forward(x), float)
+
+def test_ergodic_forward():
+    """Forward algorithm on an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    hmm.fit(X)
+    assert isinstance(hmm.forward(x), float)
+
+def test_strict_left_right_forward():
+    """Forward algorithm on a strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    hmm.fit(X)
+    assert isinstance(hmm.forward(x), float)
+
+# ======================= #
+# GMMHMM.label (property) #
+# ======================= #
+
+def test_label():
+    assert deepcopy(hmm_lr).label == 'c1'
+
+# ========================== #
+# GMMHMM.n_states (property) #
+# ========================== #
+
+def test_n_states():
+    assert deepcopy(hmm_lr).n_states == 5
 
 # ============================== #
 # GMMHMM.n_components (property) #
 # ============================== #
 
 def test_n_components():
-    assert deepcopy(hmm_diag).n_components == 5
+    assert deepcopy(hmm_lr).n_components == 1
 
-# ============================ #
-# GMMHMM.covariance (property) #
-# ============================ #
+# ================================= #
+# GMMHMM.covariance_type (property) #
+# ================================= #
 
-def test_covariance():
-    assert deepcopy(hmm_diag).covariance == 'diagonal'
+def test_covariance_type():
+    assert deepcopy(hmm_lr).covariance_type == 'full'
 
-# ================ #
-# GMMHMM.as_dict() #
-# ================ #
+# ======================== #
+# GMMHMM.n_seqs (property) #
+# ======================== #
 
-def test_as_dict_unfitted():
-    """Export an unfitted GMMHMM to dict"""
-    hmm = deepcopy(hmm_diag)
+def test_n_seqs_without_fit():
+    """Number of sequences without fitting the HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
     with pytest.raises(AttributeError) as e:
-        hmm.as_dict()
-    assert str(e.value) == 'The model needs to be fitted before it can be exported to a dict'
+        hmm.n_seqs
+    assert str(e.value) == 'The model has not been fitted and has not seen any observation sequences'
 
-def test_as_dict_fitted():
-    """Export a fitted GMMHMM to dict"""
-    hmm = deepcopy(hmm_diag)
-    hmm.set_uniform_initial()
-    hmm.set_uniform_transitions()
-    before = hmm.initial, hmm.transitions
-    with warnings.catch_warnings():
-        warnings.filterwarnings('ignore', category=DeprecationWarning)
-        hmm.fit(X)
-    d = hmm.as_dict()
+def test_n_seqs_with_fit():
+    """Number of sequences after fitting the HMM"""
+    hmm = deepcopy(hmm_lr)
+    hmm.set_random_initial()
+    hmm.set_random_transitions()
+    hmm.fit(X)
+    assert hmm.n_seqs == 3
 
-    assert d['type'] == 'GMMHMM'
-    assert d['label'] == 'c1'
-    assert d['n_states'] == 5
-    assert d['n_components'] == 5
-    assert d['covariance'] == 'diagonal'
-    assert d['topology'] == 'left-right'
-    assert_not_equal(d['model']['initial'], before[0])
-    assert_not_equal(d['model']['transitions'], before[1])
-    assert d['model']['n_seqs'] == 3
-    assert d['model']['n_features'] == 3
-    assert isinstance(d['model']['hmm'], dict)
+# ========================= #
+# GMMHMM.initial (property) #
+# ========================= #
 
-# ========== #
-# HMM.save() #
-# ========== #
+def test_initial_without_setting():
+    """Get initial state distribution without setting it"""
+    hmm = deepcopy(hmm_lr)
+    with pytest.raises(AttributeError) as e:
+        hmm.initial
+    assert str(e.value) == 'No initial state distribution has been defined'
 
-def test_save_directory():
-    """Save a GMMHMM into a directory"""
-    hmm = deepcopy(hmm_diag)
-    hmm.set_uniform_initial()
-    hmm.set_uniform_transitions()
-    with warnings.catch_warnings():
-        warnings.filterwarnings('ignore', category=DeprecationWarning)
-        hmm.fit(X)
-    try:
-        with pytest.raises(IsADirectoryError) as e:
-            hmm.save('.')
-        assert str(e.value) == "[Errno 21] Is a directory: '.'"
-    except ValueError:
-        pass
+# ============================= #
+# GMMHMM.transitions (property) #
+# ============================= #
 
-def test_save_no_extension():
-    """Save a GMMHMM into a file without an extension"""
-    try:
-        hmm = deepcopy(hmm_diag)
-        hmm.set_uniform_initial()
-        hmm.set_uniform_transitions()
-        with warnings.catch_warnings():
-            warnings.filterwarnings('ignore', category=DeprecationWarning)
-            hmm.fit(X)
-        try:
-            hmm.save('test')
-            assert os.path.isfile('test')
-        except ValueError:
-            pass
-    finally:
-        try:
-            os.remove('test')
-        except OSError:
-            pass
+def test_transitions_without_setting():
+    """Get transition matrix without setting it"""
+    hmm = deepcopy(hmm_lr)
+    with pytest.raises(AttributeError) as e:
+        hmm.transitions
+    assert str(e.value) == 'No transition matrix has been defined'
 
-def test_save_with_extension():
-    """Save a GMMHMM into a file with a .json extension"""
-    try:
-        hmm = deepcopy(hmm_diag)
-        hmm.set_uniform_initial()
-        hmm.set_uniform_transitions()
-        with warnings.catch_warnings():
-            warnings.filterwarnings('ignore', category=DeprecationWarning)
-            hmm.fit(X)
-        try:
-            hmm.save('test.json')
-            assert os.path.isfile('test.json')
-        except ValueError:
-            pass
-    finally:
-        try:
-            os.remove('test.json')
-        except OSError:
-            pass
+# ======================= #
+# GMMHMM.initial (setter) #
+# ======================= #
 
-# ============= #
-# GMMHMM.load() #
-# ============= #
+def test_left_right_initial_left_right():
+    """Set an initial state distribution generated by a left-right topology on a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    topology = _LeftRightTopology(n_states=5, random_state=rng)
+    initial = topology.random_initial()
+    hmm.initial = initial
+    assert_equal(hmm.initial, initial)
 
-def test_load_invalid_dict():
-    """Load a GMMHMM from an invalid dict"""
-    with pytest.raises(KeyError) as e:
-        GMMHMM.load({})
+def test_left_right_initial_ergodic():
+    """Set an initial state distribution generated by a left-right topology on an ergodic HMM"""
+    hmm = deepcopy(hmm_lr)
+    topology = _ErgodicTopology(n_states=5, random_state=rng)
+    initial = topology.random_initial()
+    hmm.initial = initial
+    assert_equal(hmm.initial, initial)
 
-def test_load_dict():
-    """Load a GMMHMM from a valid dict"""
-    hmm = deepcopy(hmm_diag)
-    hmm.set_uniform_initial()
-    hmm.set_uniform_transitions()
-    before = hmm.initial, hmm.transitions
-    with warnings.catch_warnings():
-        warnings.filterwarnings('ignore', category=DeprecationWarning)
-        hmm.fit(X)
-    hmm = GMMHMM.load(hmm.as_dict())
+def test_left_right_initial_strict_left_right():
+    """Set an initial state distribution generated by a left-right topology on an strict left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    topology = _StrictLeftRightTopology(n_states=5, random_state=rng)
+    initial = topology.random_initial()
+    hmm.initial = initial
+    assert_equal(hmm.initial, initial)
 
-    assert isinstance(hmm, GMMHMM)
-    assert hmm._label == 'c1'
-    assert hmm._n_states == 5
-    assert hmm._n_components == 5
-    assert hmm._covariance == 'diagonal'
-    assert isinstance(hmm._topology, _LeftRightTopology)
-    assert_not_equal(hmm._initial, before[0])
-    assert_not_equal(hmm._transitions, before[1])
-    assert hmm._n_seqs == 3
-    assert hmm._n_features == 3
-    assert isinstance(hmm._model, pg.HiddenMarkovModel)
+def test_ergodic_initial_left_right():
+    """Set an initial state distribution generated by an ergodic topology on a left-right HMM"""
+    hmm = deepcopy(hmm_e)
+    topology = _LeftRightTopology(n_states=5, random_state=rng)
+    initial = topology.random_initial()
+    hmm.initial = initial
+    assert_equal(hmm.initial, initial)
 
-def test_load_invalid_path():
-    """Load a GMMHMM from a directory"""
-    with pytest.raises(IsADirectoryError) as e:
-        GMMHMM.load('.')
+def test_ergodic_initial_ergodic():
+    """Set an initial state distribution generated by an ergodic topology on an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    topology = _ErgodicTopology(n_states=5, random_state=rng)
+    initial = topology.random_initial()
+    hmm.initial = initial
+    assert_equal(hmm.initial, initial)
 
-def test_load_inexistent_path():
-    """Load a GMMHMM from an inexistent path"""
-    with pytest.raises(FileNotFoundError) as e:
-        GMMHMM.load('test')
+def test_ergodic_initial_strict_left_right():
+    """Set an initial state distribution generated by an ergodic topology on a strict left-right HMM"""
+    hmm = deepcopy(hmm_e)
+    topology = _StrictLeftRightTopology(n_states=5, random_state=rng)
+    initial = topology.random_initial()
+    hmm.initial = initial
+    assert_equal(hmm.initial, initial)
 
-def test_load_invalid_format():
-    """Load a GMMHMM from an illegally formatted file"""
-    try:
-        with open('test', 'w') as f:
-            f.write('illegal')
-        with pytest.raises(json.decoder.JSONDecodeError) as e:
-            GMMHMM.load('test')
-    finally:
-        os.remove('test')
+def test_strict_left_right_initial_left_right():
+    """Set an initial state distribution generated by a strict left-right topology on a left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    topology = _LeftRightTopology(n_states=5, random_state=rng)
+    initial = topology.random_initial()
+    hmm.initial = initial
+    assert_equal(hmm.initial, initial)
 
-def test_load_invalid_json():
-    """Load a GMMHMM from an invalid JSON file"""
-    try:
-        with open('test', 'w') as f:
-            f.write("{}")
-        with pytest.raises(KeyError) as e:
-            GMMHMM.load('test')
-    finally:
-        os.remove('test')
+def test_strict_left_right_initial_ergodic():
+    """Set an initial state distribution generated by a strict left-right topology on an ergodic HMM"""
+    hmm = deepcopy(hmm_slr)
+    topology = _ErgodicTopology(n_states=5, random_state=rng)
+    initial = topology.random_initial()
+    hmm.initial = initial
+    assert_equal(hmm.initial, initial)
 
-def test_load_path():
-    """Load a GMMHMM from a valid JSON file"""
-    try:
-        hmm = deepcopy(hmm_diag)
-        hmm.set_uniform_initial()
-        hmm.set_uniform_transitions()
-        before = hmm.initial, hmm.transitions
-        with warnings.catch_warnings():
-            warnings.filterwarnings('ignore', category=DeprecationWarning)
-            hmm.fit(X)
-        hmm.save('test')
-        hmm = GMMHMM.load('test')
+def test_strict_left_right_initial_strict_left_right():
+    """Set an initial state distribution generated by a strict left-right topology on an strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    topology = _StrictLeftRightTopology(n_states=5, random_state=rng)
+    initial = topology.random_initial()
+    hmm.initial = initial
+    assert_equal(hmm.initial, initial)
 
-        assert isinstance(hmm, GMMHMM)
-        assert hmm._label == 'c1'
-        assert hmm._n_states == 5
-        assert hmm._n_components == 5
-        assert isinstance(hmm._topology, _LeftRightTopology)
-        assert hmm._covariance == 'diagonal'
-        assert_not_equal(hmm._initial, before[0])
-        assert_not_equal(hmm._transitions, before[1])
-        assert hmm._n_seqs == 3
-        assert hmm._n_features == 3
-        assert isinstance(hmm._model, pg.HiddenMarkovModel)
-    finally:
-        os.remove('test')
+# =========================== #
+# GMMHMM.transitions (setter) #
+# =========================== #
+
+def test_left_right_transitions_left_right():
+    """Set a transition matrix generated by a left-right topology on a left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    topology = _LeftRightTopology(n_states=5, random_state=rng)
+    transitions = topology.random_transitions()
+    hmm.transitions = transitions
+    assert_equal(hmm.transitions, transitions)
+
+def test_left_right_transitions_ergodic():
+    """Set a transition matrix generated by a left-right topology on an ergodic HMM"""
+    hmm = deepcopy(hmm_lr)
+    topology = _ErgodicTopology(n_states=5, random_state=rng)
+    transitions = topology.random_transitions()
+    with pytest.raises(ValueError) as e:
+        hmm.transitions = transitions
+    assert str(e.value) == 'Left-right transition matrix must be upper-triangular'
+
+def test_left_right_transitions_strict_left_right():
+    """Set a transition matrix generated by a left-right topology on a strict left-right HMM"""
+    hmm = deepcopy(hmm_lr)
+    topology = _StrictLeftRightTopology(n_states=5, random_state=rng)
+    transitions = topology.random_transitions()
+    hmm.transitions = transitions
+    assert_equal(hmm.transitions, transitions)
+
+def test_ergodic_transitions_left_right():
+    """Set a transition matrix generated by an ergodic topology on a left-right HMM"""
+    hmm = deepcopy(hmm_e)
+    topology = _LeftRightTopology(n_states=5, random_state=rng)
+    transitions = topology.random_transitions()
+    with pytest.warns(UserWarning) as w:
+        hmm.transitions = transitions
+    assert w[0].message.args[0] == 'Zero probabilities in ergodic transition matrix - these transition probabilities will not be learned'
+    assert_equal(hmm.transitions, transitions)
+
+def test_ergodic_transitions_ergodic():
+    """Set a transition matrix generated by an ergodic topology on an ergodic HMM"""
+    hmm = deepcopy(hmm_e)
+    topology = _ErgodicTopology(n_states=5, random_state=rng)
+    transitions = topology.random_transitions()
+    hmm.transitions = transitions
+    assert_equal(hmm.transitions, transitions)
+
+def test_ergodic_transitions_strict_left_right():
+    """Set a transition matrix generated by an ergodic topology on a strict left-right HMM"""
+    hmm = deepcopy(hmm_e)
+    topology = _StrictLeftRightTopology(n_states=5, random_state=rng)
+    transitions = topology.random_transitions()
+    with pytest.warns(UserWarning) as w:
+        hmm.transitions = transitions
+    assert w[0].message.args[0] == 'Zero probabilities in ergodic transition matrix - these transition probabilities will not be learned'
+    assert_equal(hmm.transitions, transitions)
+
+def test_strict_left_right_transitions_left_right():
+    """Set a transition matrix generated by a strict left-right topology on a left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    topology = _LeftRightTopology(n_states=5, random_state=rng)
+    transitions = topology.random_transitions()
+    with pytest.raises(ValueError) as e:
+        hmm.transitions = transitions
+    assert str(e.value) == 'Strict left-right transition matrix must only consist of a diagonal and upper diagonal'
+
+def test_strict_left_right_transitions_ergodic():
+    """Set a transition matrix generated by a strict left-right topology on an ergodic HMM"""
+    hmm = deepcopy(hmm_slr)
+    topology = _ErgodicTopology(n_states=5, random_state=rng)
+    transitions = topology.random_transitions()
+    with pytest.raises(ValueError) as e:
+        hmm.transitions = transitions
+    assert str(e.value) == 'Left-right transition matrix must be upper-triangular'
+
+def test_strict_left_right_transitions_strict_left_right():
+    """Set a transition matrix generated by a strict left-right topology on a strict left-right HMM"""
+    hmm = deepcopy(hmm_slr)
+    topology = _StrictLeftRightTopology(n_states=5, random_state=rng)
+    transitions = topology.random_transitions()
+    hmm.transitions = transitions
+    assert_equal(hmm.transitions, transitions)
+
+# ============================== #
+# GMMHMM (serialize/deserialize) #
+# ============================== #
+
+#
